Encapsulate the Device Model
This commit adds the DeviceModelTemplate abstract class for
encapsulating methods to calculate the hydrodynamics and forces
parameters for the control calculations of the generic device. The
concrete implementation for RM3 is provided in the RM3DeviceModel class.

DeviceModelTemplate may also require other functions, such as the
different control functions (ps, etc) to be encapsulated as abstract
methods as well.