Create Device Model Class Folders
Device models are now defined in class folders with the abstract methods
added as functions (although they do require obj as the first input
still). All control functions were also encapsulated into the
DeviceModelTemplate base class. As we uncover more generic code it is
likely to be placed within this class.

Models are now stored in the WecOptLib.models package.