Add Device Model Class Folders (#54)
* Create Device Model Class Folders

Device models are now defined in class folders with the abstract methods
added as functions (although they do require obj as the first input
still). All control functions were also encapsulated into the
DeviceModelTemplate base class. As we uncover more generic code it is
likely to be placed within this class.

Models are now stored in the WecOptLib.models package.

* Move subSampleFreqs to utils package and remove volatile package

* Changed header to reflect changes

Co-authored-by: ssolson <ssolson@users.noreply.github.com>