Break-out the Pseudo Spectral Code
The code for the RM3 pseudo spectral controller has been broken out from
RM3Device.psuedoSpectral into a new package called RM3, in the models
package. RM3DeviceModel was also renamed as DeviceModel and placed in
the RM3 package.

Two new helper functions are added to the RM3 package, getPSCoefficients
and getPSPhasePower which contain the work done before and during the
phase loop.