Return Power Per Frequency From Control Optimisers
This commit changes the output of the control optimisers from total
power to power per frequency plus the frequency bins used in the
calculations.

This information is stored in the `etc` output of
`DeviceModelTemplate.getPower` as `etc.powPerFreq` and `etc.freq`. The
total power output `pow` is also now returned as a positive value, which
is then turned negative when used in the optimisation of the `run`
function.

The `WecOptLib.sandbox.plot` function now demonstrates how to create a
plot for a single sea-state input. It is not clear how this scales to
multiple sea-states, particularly as the frequencies may differ between
them.