PS: Frequency vs. Power (#55)
* Fixes bug where the random phase realization was being ignored. Further adds a section which attepts to find frequency vs. 'energy' for todays discussion. This uses equation 6.11 for Bacelli 2014 Dissertation. Currently the power is not the same as caclulated using Pt and tkp. Currently I believe this may be due to scaling? by considering the uT calculation on line 154. Reaching out to @bacelli for a meeting tomorrow to discuss.

* Break-out the Pseudo Spectral Code

The code for the RM3 pseudo spectral controller has been broken out from
RM3Device.psuedoSpectral into a new package called RM3, in the models
package. RM3DeviceModel was also renamed as DeviceModel and placed in
the RM3 package.

Two new helper functions are added to the RM3 package, getPSCoefficients
and getPSPhasePower which contain the work done before and during the
phase loop.

* Add Plot Function for Investigation

This is a play function for testing plotting with the pseudo spectral
controller. It is called as WecOptLib.sandbox.plot()

* Add isClose function for comparing floats

* Remove time from energy equation and scale to get power per frequency
for pseudo spectral controller

* Fix typo

* Code clean up, comments. Change to output totalpower  and power per frequency to match PS output.

* Description in header

* Return Power Per Frequency From Control Optimisers

This commit changes the output of the control optimisers from total
power to power per frequency plus the frequency bins used in the
calculations.

This information is stored in the `etc` output of
`DeviceModelTemplate.getPower` as `etc.powPerFreq` and `etc.freq`. The
total power output `pow` is also now returned as a positive value, which
is then turned negative when used in the optimisation of the `run`
function.

The `WecOptLib.sandbox.plot` function now demonstrates how to create a
plot for a single sea-state input. It is not clear how this scales to
multiple sea-states, particularly as the frequencies may differ between
them.

* Fix docstrings

* Example sea-states data. Contains 8 Bretchnider spectrums of varying Hm0s, Tps, Nbins, and range. Running the new plot function on this sea-states produced a reasonable result and no errors.

* Added plotting as method of WecOptTool with example.

* Plot All Spectra Used in Study

The power output for all spectra used in a Study are now plotted by the
WecOptTool.plot function. The multi-spectra example data was given an
access method (WecOptLib.tests.data.example8Spectra) and the warnings it
generates were suppressed.

* Fix the example integration test to suppress plotting

* Delete sandbox plot function

* Fixes bug where the random phase realization was being ignored. Further adds a section which attepts to find frequency vs. 'energy' for todays discussion. This uses equation 6.11 for Bacelli 2014 Dissertation. Currently the power is not the same as caclulated using Pt and tkp. Currently I believe this may be due to scaling? by considering the uT calculation on line 154. Reaching out to @bacelli for a meeting tomorrow to discuss.

* Break-out the Pseudo Spectral Code

The code for the RM3 pseudo spectral controller has been broken out from
RM3Device.psuedoSpectral into a new package called RM3, in the models
package. RM3DeviceModel was also renamed as DeviceModel and placed in
the RM3 package.

Two new helper functions are added to the RM3 package, getPSCoefficients
and getPSPhasePower which contain the work done before and during the
phase loop.

* Add Plot Function for Investigation

This is a play function for testing plotting with the pseudo spectral
controller. It is called as WecOptLib.sandbox.plot()

* Add isClose function for comparing floats

* Remove time from energy equation and scale to get power per frequency
for pseudo spectral controller

* Fix typo

* Code clean up, comments. Change to output totalpower  and power per frequency to match PS output.

* Description in header

* Return Power Per Frequency From Control Optimisers

This commit changes the output of the control optimisers from total
power to power per frequency plus the frequency bins used in the
calculations.

This information is stored in the `etc` output of
`DeviceModelTemplate.getPower` as `etc.powPerFreq` and `etc.freq`. The
total power output `pow` is also now returned as a positive value, which
is then turned negative when used in the optimisation of the `run`
function.

The `WecOptLib.sandbox.plot` function now demonstrates how to create a
plot for a single sea-state input. It is not clear how this scales to
multiple sea-states, particularly as the frequencies may differ between
them.

* Fix docstrings

* Example sea-states data. Contains 8 Bretchnider spectrums of varying Hm0s, Tps, Nbins, and range. Running the new plot function on this sea-states produced a reasonable result and no errors.

* Added plotting as method of WecOptTool with example.

* Plot All Spectra Used in Study

The power output for all spectra used in a Study are now plotted by the
WecOptTool.plot function. The multi-spectra example data was given an
access method (WecOptLib.tests.data.example8Spectra) and the warnings it
generates were suppressed.

* Fix the example integration test to suppress plotting

* Delete sandbox plot function

* Improve sea-state weighting checks

* This moves common impedance calculations from cc and damping functions into the motion calculation. Added comments throughout. Moved the definintion of the RM3 or motion struct to the end of getMotion to make the code more readable.

* Typo in function def

* Adding axis labels, grid, and legend

* Adding reference to linear algebra.

Co-authored-by: Mathew Topper <damm_horse@yahoo.co.uk>