Allow Reuse of Existing NEMOH Solutions
This commit allows the reuse of existing NEMOH solutions (mainly for
plotting) rather than having to recalculate. To do this the etc
struct in getHyrodynamics is serialized into the NEMOH run directories
in the study directory. These are then reread after the optimisation
to find the directory which matches the optimal solution.

Various changes to improve the readability of the code have also been
added.