Add Types and Multiple Sea-States
The four defined data types are added to this commit (SeaState, Hydro,
Motion and Performance), which are based upon the abstract Data class.
The type classes enforce inclusion of certain parameters when
initialized (from a struct) and will store any additional parameters
given. They can also be initialized from the struct array. As in
SeaState, the type classes can also be used for data preparation, as
required. The required user defined functions must also return one of
these types.

Multiple sea-states are now solved by DefaultDevice.simulate and the
Motion and Performance object created for each are stored as an array
(in the motions and performances properties). The user can now add
a optional additional callback to DefaultBlueprint that will take the
SeaState object and all of the results in DefaultDevice and any
output from this will be added to the aggregation property of
DefaultDevice. This could also be added within the user defined cost
functions, but this allows a neater method for repeated use.