Begin division between using NEMOH for meshing versus BEM.
Requires two calls to the mesh command, annoyingly, one to mesh and
one for the hydrostatic calculations. Also if more complex meshes are
passed to mesh for the hydrostatic calculations it might fail, as it
says there is an upper limit of 2000 panels. More complex meshes might
need to be automatically simplified for the hydrostatic calculations
to be obtainable.