Add Optional Model Parameter Passing
This commit adds optional passing of parameters to the model
callbacks. The inputs to Blueprint.makeDevices are now struct arrays
with the type and parameters for each desired geometry, controller
or model (parameters) given in the type and params fields. The values
passed to the params fields should be cell arrays.