Blueprint Model (#144)
* Add Blueprint and derived Device classes to replace the existing 
RM3Study class functionality.

* Callbacks are used to allow the user to implement strategy pattern for 
Blueprint subclasses.

* Blueprint.makeDevices creates Device objects and takes struct arrays 
as input with the type and parameters for each desired geometry, 
controller or model (parameters) given in the type and params fields. The 
values passed to the params fields should be cell arrays. 

* Devices store the options used to create them, and the SeaState used in 
the last call to simulate, in their attributes. This can be used to search 
for a particular simulation after a call to the recoverDevices method
which rebuilds all devices associated to a Blueprint. 

* Four data types classes are added to this commit (SeaState, Hydro,
Motion and Performance), which are based upon the abstract Data class.
The type classes enforce inclusion of certain parameters when
initialized (from a struct) and will store any additional parameters
given. They can also be initialized from the struct array. As in
SeaState, the type classes can also be used for data preparation, as
required. The required user defined functions must also return one of
these types. Initialization of type classes with struct arrays uses the 
"typeArray" function. The validateArray method can be used to finish
work on arrays (i.e. that require access to the entire array of data).

* Serialization of devices simulations is re-implemented, with device 
objects for each call of simulate being stored in a folder within a parent 
folder created by the Blueprint. The recoverDevices method is added to 
Blueprint which will reload all of the stored simulations into an Array of 
Device objects. 

* AutoFolder subclasses that are created using a parent directory are not 
deleted on destruction. Destruction of the parent object will clear them 
up. 

* The powerPerFreq plot is reimplemented using a Device object as input. 

* AxiMesh and NEMOH classes are added as concrete Mesher and Solver 
classes respectively.
 
* The writing of the Nemoh.cal file is improved using templates rather than 
multiple fprintf statements.

* Shortcut functions types, mesh and solver added to the top level to provide 
functional access to main calls of the underlying classes.

* Optional model parameter passing adds optional passing of parameters to 
the model callbacks.  

* Added RM3 example port

* Added beginnings of WaveBot model port.

* Repurposed example docs to show an example of optimizing an existing
WEC model and added a section about creating a new WEC model.

* Updated inline docs and API page.