Reduce to structured example with library functions. (#161)
This commit changes the examples to use a pair of barebones function calls with 
master scripts and restructures the WecOptTool toolbox as a library to support 
the development of the examples. This involved the removal of the Blueprint, 
Device and Data classes, and promotion of the SeaState and AutoFolder classes 
to the top level namespace. The WecOptLib namespace has also been removed. 

This commit also adds the first single body example "WaveBot".

Other changes are:

* tests have been moved to a top level directory

* all copyright statements are now located correctly within the files

* AutoFolder now deals with storing intermediate results from optimizations

* TempFolder base class added for access to a temporary folder that will
  not automatically delete itself

* Added weightings to example spectra

Co-authored-by: Ryan Coe <rcoe@sandia.gov>