Update Dependency Description and Checker
This commit updates the descriptions of the dependencies in the
README, to distinguish between those required for the toolbox and the
examples.

Additionally the global optimization toolbox , a new dependency for the
examples, is added and dependencyCheck was modified to look for it.

The hasParallelToolbox function was generalized to look for any
toolbox (now hasToolbox), as was printing of dependencies in
dependencyCheck.