Fix details not being unrolled when clicking on summary in Chrome (#181)
Chrome does not treat <a> tags inside a <summary> tag the same way as Firefox, such that the details tag will not unroll when the summary is clicked. This commit swaps the <a> tags for a custom tag <clickme> that colours the summaries like links, without being links. Chrome now acts as expected.

Couple of other small fixes to docstrings were included.