Add note attribute to SeaState class
This commit adds the note attribute to the SeaState class to carry
a description of each spectrum. This allows direct definition when
initializing from WAFO for instance. If the note field is not set
in the input struct then a default value is added.

A test was also added to check the plotting code shown in the
RM3 optimization.m example in the online docs.