Added "Citing WecOptTool" and "Resources" pages (#1)
This commit adds two new pages, one for storing a single citation
for WecOptTool in a textual and bibtex form, ala Numpy. The
second page includes additional resources for the user. Here,
the webinar slides are embedded in a viewer, and some additional
links for further reading are also included.

Numbering was also added to the file names to ensure that the
cited references get processed and numbered in the correct order.

The build requirements are also updated in the README and travisCI
config.