Calculate standard error on PS results
This commit uses the central value theorem to facilitate calculation
of the error on the mean results of the PS controller, using the
standard error from a normal distribution.

Two new functions are added to the WecOptTool.math package called
standardErrorMeasure and standardErrorReduce, the first of which
measures the standard error after a given number of samples and the
second of which keeps sampling until a desired level of error is
found.

An example script to show how the standard error varies with the
stroke of the RM3 device using the PS controller is added to the
RM3 example directory called PSError.m