Implement option for maximum samples with standardError in reduce mode
Adds the option 'maxN' to standardError to control the maximum number
of samples when in reduce mode. standardError now always returns both
the error and the number of samples in its results.

The option 'onError' is also added to control behavior when maxN is
exceeded. In 'warn' mode a warning will be issued and in 'raise' mode
an error is raised.